% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getOffspringTip}
\alias{getOffspringTip}
\title{Offspring tips of inner node}
\usage{
getOffspringTip(node, tree)
}
\arguments{
\item{node}{single integer value corresponding to the index of a node from \code{tree}}

\item{tree}{rooted tree object of class phylo}
}
\value{
Integer vector with all indices of offspring tips of \code{node} in \code{tree}.
}
\description{
Returns all tips of a \code{tree} which are an offspring of 
a given \code{node} of the tree
}
\examples{
library(ape)
n <- 2 * 50
tree <- rtree(n, rooted = TRUE)
aktiveNode <- n + 50
offAktiveNode <- getOffspringTip(aktiveNode,tree)
}
