% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeSeg.R
\name{treeSeg}
\alias{treeSeg}
\title{Segmentation of tip-observations from tree structure}
\usage{
treeSeg(y, tree, q, alpha, fam, tipOrder, lengths, checkOrder = TRUE)
}
\arguments{
\item{y}{observations, binary if \code{fam} equals "binomial" and numeric if \code{fam} equals "gauss". 
Missing data should be NA's for which predictions are computed.}

\item{tree}{rooted tree object with tip labels \code{1:length(y)} of class phylo. 
The tip labels are assumed to be of consecutive order, such that for any inner node the offspring tips are of the form i, i+1, ..., j-1, j.}

\item{q}{threshold parameter corresponding to \code{1-alpha} quantile of multiscale statistic}

\item{alpha}{confidence level (if specified \code{q} is silently ignored)}

\item{fam}{specifies distribution of data and can take either of the values "gauss" or "binomial" (default) where the letter assumes n=1, 
i.e. Bernoulli distribution}

\item{tipOrder}{specifies ordering of tips and can take either of the values "cladewise" (default) or "unchanged". 
The former applies \code{ape::reorder.phylo(tree)} prior to analysis, to make outcome independent of tip ordering.}

\item{lengths}{length of intervals considered; can take either of the value "all" or "dyadic"; by default up to a sample size of 500 all lengths, otherwise only dyadic lengths, which gives a significant seep up in computation time.}

\item{checkOrder}{\code{\link{logical}} (TRUE by default), indicates whether consecutive order of tip nodes should be checked.}
}
\value{
A list with values \code{numbAN}, \code{mlAN}, \code{mlP}, \code{confSetAN}, 
and \code{confBandP}.

\item{numbAN}{single integer value which contains the number of estimated active nodes constituting a \code{1-alpha} lower confidence bound}
\item{mlAN}{vector of integer values containing the estimated active nodes (maximum likelihood solution under multiscale constraint)}
\item{mlP}{vector of numeric values between 0 and 1 with estimated success probabilities for tips when \code{fam = "binomial"} and vector of numeric values with estimated observational means for tips when \code{fam = "gauss"} }
\item{confSetAN}{vector of integer values constituting \code{1-alpha} confidence set for active nodes}
\item{confBandP}{matrix with two columns and number of rows equal to \code{length(y)} constituting \code{1-alpha} confidence band for success probabilites and mean values, respectively, of observations \code{y}}
}
\description{
For a given tree with observations at the tips, 
performs segmentation which corresponds to a set of inner nodes in the tree. 
The method is based on a multiscale approach. This controls the probability of 
overestimating the number of active nodes at level alpha and yields confidence statements 
for the set of active nodes and the observational means (success probabilities for binary data).
}
\note{
If neither \code{q} nor \code{alpha} is specified, \code{alpha = 0.1} is selected.
}
\examples{
library(ape)
n <- 2 * 50
tree <- rtree(n, rooted = TRUE)
aktiveNode <- n + 50
offAktiveNode <- getOffspringTip(aktiveNode,tree)
prob1 <- 0.1
prob2 <- 0.75
p <- prob1 + prob2 * is.element(1:n, offAktiveNode)
y <- rbinom(n, 1, p)
ans <- treeSeg(y, tree, alpha = 0.1)
yGauss <- rnorm(n, p, sd = 0.1)
ansGauss <- treeSeg(yGauss, tree, alpha = 0.1, fam = "gauss")
}
\references{
Behr, M., Ansari, M. A., Munk, A., Holmes, C. (2020)
Testing for dependence on tree structures.
bioRxiv:622811.

Frick, K., Munk, A., Sieling, H. (2014) 
Multiscale change-point inference. 
With discussion and rejoinder by the authors. 
Journal of the Royal Statistical Society, Series B 76(3), 495:580.
}
\seealso{
\code{\link{getOffspringTip}}, \code{\link{treeTest}}
}
