% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeTest.R
\name{treeTest}
\alias{treeTest}
\title{Multiscale test for association between tips and tree structure}
\usage{
treeTest(y, tree, q, alpha, fam, tipOrder)
}
\arguments{
\item{y}{observations, binary if \code{fam} equals "binomial" and numeric if \code{fam} equals "gauss". Missing data should be NA's and is silently ignored.}

\item{tree}{rooted tree object with tip labels \code{1:length(y)} of class phylo. 
The tip labels are assumed to be of consecutive order, such that for any inner node the offspring tips are of the form i, i+1, ..., j-1, j.
If missing, it is assumed that \code{y} is already ordered with respect to the tree-tips.}

\item{q}{threshold parameter corresponding to \code{1-alpha} quantile of multiscale statistic}

\item{alpha}{confidence level (if specified \code{q} is silently ignored)}

\item{fam}{specifies distribution of data and can take either of the values "gauss" or "binomial" (default) where the letter assumes n=1, i.e. Bernoulli distribution}

\item{tipOrder}{specifies ordering of tips and can take either of the values "cladewise" (default) or "unchanged". 
The former applies \code{ape::reorder.phylo(tree)} prior to analysis, to make outcome independent of tip ordering.}
}
\value{
A single integer number, where 0 corresponds to acceptance of null hypothesis at given level \code{alpha}. 
Any integer lager than 0 corresponds to rejection of null hypothesis at given level \code{alpha} 
and gives a lower bound on the minimal number of active nodes in the full \code{\link{treeSeg}} segmentation.
}
\description{
For a given tree with observations at the tips, 
tests for association between the tree structure and the observational distribution.
The method is based on a multiscale approach. The null hypothesis assumes no association, which corresponds to
i.i.d. tree-ordered observations. 
In general, this is computationally faster than deriving the full \code{\link{treeSeg}} segmentation.
}
\note{
If neither \code{q} nor \code{alpha} is specified, \code{alpha = 0.1} is selected.
}
\examples{
library(ape)
n <- 2 * 50
tree <- rtree(n, rooted = TRUE)
aktiveNode <- n + 50
offAktiveNode <- getOffspringTip(aktiveNode,tree)
prob1 <- 0.1
prob2 <- 0.75
p <- prob1 + prob2 * is.element(1:n, offAktiveNode)
y <- rbinom(n, 1, p)
ans <- treeTest(y, tree, alpha = 0.1)
yGauss <- rnorm(n, p, sd = 0.1)
ansGauss <- treeTest(yGauss, tree, alpha = 0.1, fam = "gauss")
}
\references{
Behr, M., Ansari, M. A., Munk, A., Holmes, C. (2020)
Testing for dependence on tree structures.
bioRxiv:622811 

Frick, K., Munk, A., Sieling, H. (2014) 
Multiscale change-point inference. 
With discussion and rejoinder by the authors. 
Journal of the Royal Statistical Society, Series B 76(3), 495:580.
}
\seealso{
\code{\link{treeSeg}}
}
