% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getOffspringTip}
\alias{getOffspringTip}
\title{Offspring tips of given inner node}
\usage{
getOffspringTip(node, tree)
}
\arguments{
\item{node}{single integer value which contrains the index of a node from \code{tree}}

\item{tree}{rooted tree object}
}
\value{
Integer vector with all indeces of offspring tips of \code{node} in \code{tree}.
}
\description{
Gives back all tips of a \code{tree} which are an offspring of 
a given \code{node} of the tree
}
\examples{
library(ape)
n <- 2 * 50
tree <- rtree(n, rooted = TRUE)
aktiveNode <- n + 50
offAktiveNode <- getOffspringTip(aktiveNode,tree)
}
