% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{offspringAll}
\alias{offspringAll}
\title{All offspring nodes of inner node}
\usage{
offspringAll(node, tree)
}
\arguments{
\item{node}{single integer value corresponding to the index of a node from \code{tree}}

\item{tree}{rooted tree object}
}
\value{
Integer vector with all indeces of offspring nodes of \code{node} in \code{tree}.
}
\description{
Returns all nodes of a \code{tree} which are an offspring of 
a given \code{node} of the tree
}
\examples{
library(ape)
n <- 2 * 50
tree <- rtree(n, rooted = TRUE)
aktiveNode <- n + 50
offAktiveNode <- offspringAll(aktiveNode,tree)
}
